stock-dashboard-backend/requirements.txt
langchain_core
langgraph
langchain_openai
python-dotenv